% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\title{Get the dataset associated with a ticket_id}
\usage{
get_dataset(ticket_id)
}
\arguments{
\item{ticket_id}{a 32-character ticket_id supplied by \code{get_ticket}. Please note that this is a secret value and should not be shared with anyone else.}
}
\value{
a dataframe containing the data associated with this simulation
}
\description{
Getting a dataset is the second stage in analyzing a simulation from
causalbenchmarks.org. Once you have received a ticket_id from \code{get_new_ticket}
, you can use this function to download the dataset to analyze.
Requesting a dataset stars a timer that runs until the results
of the analysis are submitted, in order to track the speed of your algorithm.
}
