% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_new_ticket.R
\name{get_new_ticket}
\alias{get_new_ticket}
\title{Request a new ticket from causalbenchmarks.org}
\usage{
get_new_ticket(algorithm_key)
}
\arguments{
\item{algorithm_key}{the 30-character algorithm key associated with the algorithm you are requesting a ticket for.
These tickets can be located in your API console at \url{causalbenchmarks.org/dashboard}}
}
\value{
a 32-character ticket_id used in \code{get_dataset}, and \code{submit_estimate}. Please note that this is a secret value and should not be shared with anyone else.
}
\description{
Requesting a ticket is the first stage of benchmarking a causal inference algorithm using causalbenchmarks org.
A ticket allows you to request the generation of a dataset and gives you a unique URL at which to submit your results
once your analysis has completed. The request_ticket function requests a new ticket given an algorithm key
(a 32-character string available in your API dashboard).
}
