% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_estimate.R
\name{submit_estimate}
\alias{submit_estimate}
\title{Request a new ticket from causalbenchmarks.org}
\usage{
submit_estimate(ticket_id, estimate, ci_lo = NULL, ci_hi = NULL)
}
\arguments{
\item{ticket_id}{a 32-character ticket_id supplied by \code{get_ticket}. Please note that this is a secret value and should not be shared with anyone else.}
}
\description{
Requesting a ticket is the first stage of benchmarking a causal inference algorithm using causalbenchmarks org.
A ticket allows you to request the generation of a dataset and gives you a unique URL at which to submit your results
once your analysis has completed. The request_ticket function requests a new ticket given an algorithm key
(a 32-character string available in your API dashboard).
}
